1. Entrance 씬에서 로비 서버에 Join해서 로비 씬으로 이동하는 로직 구현 2. 네이밍 수정 - SetJSONObject -> SetByJSONObject 3. GameStartToC 프로토콜 구현

+++ Client_Root/Client/Assets/Scripts/Room/Games/IGame.cs
@@ -8,7 +8,7 @@ public abstract class IGame : MonoBehaviour
     [SerializeField] protected Camera               m_CameraMain = null;
     [SerializeField] protected CameraController     m_CameraController;
 
-    protected Dictionary<int, ICharacter>           m_dicCharacter = new Dictionary<int, ICharacter>();
+    protected Dictionary<ulong, ICharacter>         m_dicCharacter = new Dictionary<ulong, ICharacter>();
 
     public abstract void OnRecvMessage(IMessage msg);
 }
